encode functiont that can handle all channels
This makes several small changes to the base type of `encode`.

First, rather than stopping without a return if a `field` or `value` is missing, I provided a message and returned `vl`. This is the pattern used later when `type` is invalid, and I thought that it was better to return the vegalite object on failure rather than having no return. This ensures a functional pipe continues even if this function call has no effect.

Second, it adds `chnl` as an argument to replace the separate use of `$` based on the function name as was done with `axis`.