Adjusts all encode_* functions to use encode
Of note is a new pattern for the `order`, `path`, and `details` channels. Unlike the other channels, these functions cannot use `values` and must specify a field. Each of the encode functions for these channels maintain `field` and `value` as formals. They check for the presence of `field` and fail without it. If `value` is additionally present, `encode_*` supplies a message letting the caller know that `value` will be ignored in favor of `field.